[artf1329111]: Fix for Coverity issue in RfStateMonitor.cpp
   - Argument "old_state" to format specifier "%d" was expected to have
type "int" but has type "NfcRfState".
   - Argument "state" to format specifier "%d" was expected to have
type "int" but has type "NfcRfState".

+++ vendor_extns/libnfc_reader_vendor_extn/src/features/nci_state_monitor/RfStateMonitor.cpp
@@ -113,7 +113,7 @@ void RfStateMonitor::updateNfcRfState(NfcRfState state) {
   const NfcRfState old_state = nfcRfState;
   NXPLOG_EXTNS_D(NXPLOG_ITEM_NXP_GEN_EXTN,
                  "RfStateMonitor %s: RF state old:%d new:%d", __func__,
-                 old_state, state);
+                 static_cast<int>(old_state), static_cast<int>(state));
   nfcRfState = state;
 }
